Add initial fuzz harness, and fix a lexer bug
The lexer bug was that integer constants were being parsed as
usizes. This meant that a very large (but valid) integer token (though admittedly
the token is too large to be stored in any C type) would cause
the lexer to panic due to the token being unrepresentable as a
usize.